Split annotations into utils/features/supervision
Closes #92.
