+++ fonduer/features/__init__.py
@@ -1,0 +1,3 @@
+from fonduer.features.feature_annotator import FeatureAnnotator
+
+__all__ = ["FeatureAnnotator"]
