+++ fonduer/supervision/__init__.py
@@ -1,7 +1,3 @@
-from fonduer.supervision.annotations import (
-    FeatureAnnotator,
-    LabelAnnotator,
-    load_gold_labels,
-)
+from fonduer.supervision.label_annotator import LabelAnnotator, load_gold_labels
 
-__all__ = ["load_gold_labels", "FeatureAnnotator", "LabelAnnotator"]
+__all__ = ["load_gold_labels", "LabelAnnotator"]
